feat: msg91 otp service integrated

+++ src/services/otpServices.ts
@@ -1,10 +1,52 @@
+import axios from "axios";
+import bcrypt from "bcrypt";
+import { randomInt } from "crypto";
+
 import Otp from "../database/models/otp";
 
+const generateOtp = () => String(randomInt(0, 1_000_000)).padStart(6, "0");
+
+const sendOtpService = async (mobile, otp) => {
+  try {
+    const data = {
+      template_id: process.env.MSG91_TEMPLATE_ID,
+      short_url: "0",
+      recipients: [
+        {
+          mobiles: "91" + mobile,
+          var1: otp,
+        },
+      ],
+    };
+
+    const headers = {
+      authkey: process.env.MSG91_AUTH_KEY,
+      "content-type": "application/json",
+    };
+
+    const res = await axios.post(process.env.MSG91_BASE_URL, data, {
+      headers,
+    });
+    return res.data;
+  } catch (err) {
+    console.error("MSG91 OTP send failed:", err.response?.data || err.message);
+    throw new Error("Failed to send OTP");
+  }
+};
+
 export const createOtp = async (mobile: string) => {
-  const otp = "123456";
+  const otp = generateOtp();
+
+  await sendOtpService(mobile, otp);
+
   const expiresAt = new Date(Date.now() + 5 * 60 * 1000);
 
-  return await Otp.create({ mobile, otp, expiresAt });
+  const hashedOtp = await bcrypt.hash(otp, 10);
+
+  await Otp.create({ mobile, otpHash: hashedOtp, expiresAt });
+  return {
+    otp,
+  };
 };
 
 export const verifyOtpFromDB = async (
@@ -13,10 +55,13 @@ export const verifyOtpFromDB = async (
   consume = true
 ): Promise<boolean> => {
   const record = await Otp.findOne({
-    where: { mobile, otp },
+    where: { mobile },
     order: [["createdAt", "DESC"]],
   });
   if (!record || record.expiresAt < new Date()) return false;
+
+  const isMatch = await bcrypt.compare(otp, record.otpHash);
+  if (!isMatch) return false;
 
   if (consume) {
     await Otp.destroy({ where: { mobile } });
